requirements.txt
django>=5.0.0,<5.1.0
djangorestframework
pyyaml
requests
django-cors-headers
